refactor: split tools/robinhood_trading_tools.py (Part 1: Stock Orders)
This is the first part of splitting the monolithic robinhood_trading_tools.py.
It moves the 9 stock order functions to src/open_stocks_mcp/tools/trading/orders_stock.py.
This reduces the diff size to under the safety threshold of 2500 lines.

- Created src/open_stocks_mcp/tools/trading/orders_stock.py
- Updated imports in app.py and robinhood.py
- Updated @patch paths in tests/unit/test_trading_tools.py
- Kept option and management tools in robinhood_trading_tools.py for now.

Unblocks PR #694
